// mvp/types.ts
export interface PaymentMethod {
  id: string;
  name: string;
  type: 'Cash' | 'Card' | 'Digital Wallet' | 'Credit' | 'Other';
  status: 'Active' | 'Inactive';
}